--- pkg/job/eventmanager.go
@@ -1,13 +1,0 @@
-// Copyright (c) Facebook, Inc. and its affiliates.
-//
-// This source code is licensed under the MIT license found in the
-// LICENSE file in the root directory of this source tree.
-
-package job
-
-// EventEmitterFetcher wraps various job event interfaces
-type EventEmitterFetcher interface {
-	RequestEmitterFetcher
-	ReportEmitterFetcher
-	InfoFetcher
-}
