+++ pkg/storage/storage.go
@@ -12,20 +12,13 @@ import (
 	"github.com/facebookincubator/contest/pkg/types"
 )
 
-// storage defines the events storage engine used by ConTest. It can be overridden
+// storage defines the storage engine used by ConTest. It can be overridden
 // via the exported function SetStorage.
 var storage Storage
 
-// Storage defines the interface that storage engines must implement
-type Storage interface {
-	// Test events storage interface
-	StoreTestEvent(event testevent.Event) error
-	GetTestEvents(eventQuery *testevent.Query) ([]testevent.Event, error)
-
-	// Framework events storage interface
-	StoreFrameworkEvent(event frameworkevent.Event) error
-	GetFrameworkEvent(eventQuery *frameworkevent.Query) ([]frameworkevent.Event, error)
-
+// JobStorage defines the interface that implements persistence for job
+// related information
+type JobStorage interface {
 	// Job request interface
 	StoreJobRequest(request *job.Request) (types.JobID, error)
 	GetJobRequest(jobID types.JobID) (*job.Request, error)
@@ -33,6 +26,19 @@ type Storage interface {
 	// Job report interface
 	StoreJobReport(report *job.JobReport) error
 	GetJobReport(jobID types.JobID) (*job.JobReport, error)
+}
+
+// Storage defines the interface that storage engines must implement
+type Storage interface {
+	JobStorage
+
+	// Test events storage interface
+	StoreTestEvent(event testevent.Event) error
+	GetTestEvents(eventQuery *testevent.Query) ([]testevent.Event, error)
+
+	// Framework events storage interface
+	StoreFrameworkEvent(event frameworkevent.Event) error
+	GetFrameworkEvent(eventQuery *frameworkevent.Query) ([]frameworkevent.Event, error)
 }
 
 // TransactionalStorage is implemented by storage backends that support transactions.
